tstore: Cleanup.
