+++ politeiad/backendv2/tstorebe/tstore/anchor.go
@@ -7,7 +7,9 @@ package tstore
 import (
 	"bytes"
 	"crypto/sha256"
+	"encoding/base64"
 	"encoding/hex"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"time"
@@ -15,6 +17,7 @@ import (
 	dcrtime "github.com/decred/dcrtime/api/v2"
 	"github.com/decred/dcrtime/merkle"
 	"github.com/decred/politeia/politeiad/backendv2/tstorebe/store"
+	"github.com/decred/politeia/util"
 	"github.com/google/trillian"
 	"github.com/google/trillian/types"
 	"google.golang.org/grpc/codes"
@@ -566,4 +569,59 @@ func (t *Tstore) anchorTrees() error {
 	go t.anchorWait(anchors, digests)
 
 	return nil
+}
+
+func convertBlobEntryFromAnchor(a anchor) (*store.BlobEntry, error) {
+	data, err := json.Marshal(a)
+	if err != nil {
+		return nil, err
+	}
+	hint, err := json.Marshal(
+		store.DataDescriptor{
+			Type:       store.DataTypeStructure,
+			Descriptor: dataDescriptorAnchor,
+		})
+	if err != nil {
+		return nil, err
+	}
+	be := store.NewBlobEntry(hint, data)
+	return &be, nil
+}
+
+func convertAnchorFromBlobEntry(be store.BlobEntry) (*anchor, error) {
+	// Decode and validate data hint
+	b, err := base64.StdEncoding.DecodeString(be.DataHint)
+	if err != nil {
+		return nil, fmt.Errorf("decode DataHint: %v", err)
+	}
+	var dd store.DataDescriptor
+	err = json.Unmarshal(b, &dd)
+	if err != nil {
+		return nil, fmt.Errorf("unmarshal DataHint: %v", err)
+	}
+	if dd.Descriptor != dataDescriptorAnchor {
+		return nil, fmt.Errorf("unexpected data descriptor: got %v, want %v",
+			dd.Descriptor, dataDescriptorAnchor)
+	}
+
+	// Decode data
+	b, err = base64.StdEncoding.DecodeString(be.Data)
+	if err != nil {
+		return nil, fmt.Errorf("decode Data: %v", err)
+	}
+	digest, err := hex.DecodeString(be.Digest)
+	if err != nil {
+		return nil, fmt.Errorf("decode digest: %v", err)
+	}
+	if !bytes.Equal(util.Digest(b), digest) {
+		return nil, fmt.Errorf("data is not coherent; got %x, want %x",
+			util.Digest(b), digest)
+	}
+	var a anchor
+	err = json.Unmarshal(b, &a)
+	if err != nil {
+		return nil, fmt.Errorf("unmarshal anchor: %v", err)
+	}
+
+	return &a, nil
 }
